requirements.txt
autogen-agentchat>=0.5.5
autogen-ext[openai]>=0.5.5
toml>=0.10.0
click>=8.1.8
faker>=0.7.0
loguru>=0.7.3
duckduckgo-search>=8.0.0
ruamel.yaml==0.18.6
gradio>5.25.0
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0.0

# Dev dependencies
pytest==8.3.5
pytest-cov==6.0.0
pytest-asyncio==0.26.0
ruff>=0.3.0

